broker: Use a probabilistic model for vacancy
The vacancy is essentially the probability of observing no other
microRNA whose footprint overlaps the footprint of the considered
position. This naturally fits the case 'k = 0' of the Poisson-Binomial
distribution.

Let 'mirbooking_broker_get_target_site_vacancy' return a probability.

+++ src/mirbooking-broker.h
@@ -53,7 +53,7 @@ gboolean               mirbooking_broker_run_finish              (MirbookingBrok
                                                                   GAsyncResult      *result,
                                                                   GError           **error);
 GArray *               mirbooking_broker_get_target_sites        (MirbookingBroker *self);
-gfloat                 mirbooking_broker_get_target_site_vacancy (MirbookingBroker           *self,
+gdouble                mirbooking_broker_get_target_site_vacancy (MirbookingBroker           *self,
                                                                   const MirbookingTargetSite *target_site);
 gdouble                mirbooking_broker_get_target_silencing    (MirbookingBroker *self,
                                                                   MirbookingTarget *target);
